Some refactoring

+++ Exercise for the ASP.NET Core Fundamentals Exam/Library/Controllers/BooksController.cs
@@ -1,5 +1,6 @@
 ﻿using Library.Services;
 using Library.View_Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
 
@@ -16,6 +17,11 @@ namespace Library.Controllers
 
         public async Task<IActionResult> All()
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.RedirectToAction("Login", "User");
+            }
+
             var model = await this.bookService.GetAllAsync();
 
             return View(model);
@@ -23,6 +29,11 @@ namespace Library.Controllers
 
         public async Task<IActionResult> Add()
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.RedirectToAction("Login","User");
+            }
+
             var addModel = new AddBookViewModel()
             {
                 Categories = await this.bookService.GetCategoryAsync(),
@@ -54,6 +65,11 @@ namespace Library.Controllers
 
         public async Task<IActionResult> AddToCollection(int bookId)
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.RedirectToAction("Login", "User");
+            }
+
             try
             {
                 var userId = this.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
@@ -70,6 +86,11 @@ namespace Library.Controllers
 
         public async Task<IActionResult> Mine()
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.RedirectToAction("Login", "User");
+            }
+
             var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
             var model = await this.bookService.GetMineAsync(userId);
 
@@ -78,6 +99,11 @@ namespace Library.Controllers
 
         public async Task<IActionResult> RemoveFromCollection(int bookId)
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.RedirectToAction("Login", "User");
+            }
+
             var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
             await this.bookService.RemoveMovieFromCollectionAsync(bookId, userId);
